+++ src/main/java/racingcar/controller/request/RacingInfoRequest.java
@@ -1,0 +1,22 @@
+package racingcar.controller.request;
+
+public class RacingInfoRequest {
+    private String names;
+    private int count;
+
+    public RacingInfoRequest() {
+    }
+
+    public RacingInfoRequest(final String names, final int count) {
+        this.names = names;
+        this.count = count;
+    }
+
+    public String getNames() {
+        return names;
+    }
+
+    public int getCount() {
+        return count;
+    }
+}
